fixed prevState issue
