--- package/react/index.ts
@@ -1,4 +1,0 @@
-import { NeuronClient } from "./NeuronClient";
-import { PrivateNeuronClient } from "./privateNeuron/PrivateNeuronClient";
-
-export { NeuronClient, PrivateNeuronClient };
