+++ tutorials/vanilla_KILL/dynamic-text.html
